requirements.txt
numpy==1.24.3
Pillow==9.4.0
Pillow==10.1.0
scikit_learn==1.3.0
torch==1.13.0
torchvision==0.14.0
tqdm==4.65.0
